refactor: remove v1alpha1 references
start removing v1alpha1 references and tests
