+++ backend/pkg/api/connect/integration/rpconnect_test.go
@@ -7,6 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+//go:build integration
+
 package integration
 
 import (
